Cleaned up or made revisions to the code files. Also deleted useless scripts
