+++ Assets/Scenes/SampleSceneDummy.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: db29b82e864db495582466c3f0116d8d
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
