--- Assets/Scripts/DisableGrab.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 4b41e27d5cdaa4459b366590cd17715d
