--- Assets/Scripts/FramingCamera.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 48962b15726ce4e61b49ac668223fa75
