--- Assets/Scripts/GlowRadius.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 812f5751736d24056a4561db363b91ab
