--- Assets/Scripts/IInteraction.cs
@@ -1,6 +1,0 @@
-using UnityEngine;
-
-public interface IInteraction
-{
-    void Interactt();
-}
