--- Assets/Scripts/IInteraction.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: e02e01e0f355c4a05833718f78442983
