--- Assets/Scripts/InteractInputListener.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: e26597b7e05134563b10c25fcd689992
