+++ Assets/Scripts/Interact_Scripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 637f8983cc3e44d10b10a757cb10b3e4
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
