--- Assets/Scripts/Interactable.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: ee94c4ccc9bc642469e201bae8fde11d
