--- Assets/Scripts/NonInteractableGroup.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 5527fd451050b47428377ab3fb6ab09c
