+++ Assets/Scripts/PickUp_Scripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f41fd3dac4a67460d973f2a86e9867d2
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
