--- Assets/Scripts/PicupItem.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 74f3a1b6540984883bf6bbb0082f3c3f
