--- Assets/Scripts/PlayerArrowController.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 94ecf579c8ac8453a9ba87205dd10d9c
