--- Assets/Scripts/PlayerInteractor.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: a249307a38cb443708befdc49c2500b5
